+++ lib/framework/widgets/tool_bar_widgets.dart
@@ -9,7 +9,7 @@ PreferredSizeWidget customAppBarWithShadow(
     Color? titleColor,
     Icon? leadingIcon,
     Color? backgroundColor,
-    List<Icon>? actionIcon, 
+    List<Icon>? actionIcon,
     }) {
   return PreferredSize(
     preferredSize: Size.fromHeight(AppBar().preferredSize.height),
@@ -84,95 +84,7 @@ Widget customBottomNavigator() {
   );
 }
 
-class CustomfloatingActionButton extends StatelessWidget {
-  const CustomfloatingActionButton({super.key});
 
-  @override
-  Widget build(BuildContext context) {
-    return FloatingActionButton(
-      onPressed: () {},
-      backgroundColor: CustomColors.buttonBackgroundCreamColor,
-      shape: const CircleBorder(),
-      elevation: Elevation.large,
-      child: IconWidget(
-        icon: Icons.add,
-        size: Sizes.size18.sp,
-      ),
-    );
-  }
-}
-
-class CustomBottomNavigationBar extends StatelessWidget {
-  const CustomBottomNavigationBar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BottomAppBar(
-      color: CustomColors.navigationBackgroundColor,
-      shape: const CircularNotchedRectangle(),
-      height: WidgetsSizes.bottomBarHeight,
-      child: Center(
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: [
-            _buildNavItem(
-              icon: Icons.home,
-              label: 'Home',
-              color: CustomColors.buttonBackgroundCreamColor,
-              isActive: true,
-            ),
-            _buildNavItem(
-              icon: Icons.explore,
-              label: 'General',
-              color: CustomColors.buttonBackgroundCreamColor,
-              isActive: false,
-            ),
-            SizedBox(
-              width: Sizes.size10.sp,
-            ),
-            _buildNavItem(
-              icon: Icons.person,
-              label: 'Profile',
-              color: CustomColors.buttonBackgroundCreamColor,
-              isActive: false,
-            ),
-            _buildNavItem(
-              icon: Icons.settings,
-              label: 'Settings',
-              color: CustomColors.buttonBackgroundCreamColor,
-              isActive: false,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildNavItem({
-    required IconData icon,
-    required String label,
-    required Color color,
-    required bool isActive,
-  }) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        IconWidget(
-          icon: icon,
-          size: Sizes.size24.sp,
-          color: isActive ? color : CustomColors.lightBackgroundColor,
-        )
-        // Spacers.widthSmall,
-        // TextWidget(
-        //   text: label,
-        //   color: isActive ? color : CustomColors.lightBackgroundColor,
-        //   // size: Sizes.size2.h,
-        //   type: isActive ? FontWeight.bold : FontWeight.normal,
-        // ),
-      ],
-    );
-  }
-}
 
 Widget curvedBoxTapable({required String text, IconData? icon, Color? color}) {
   return InkWell(
