+++ lib/framework_demo.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:pinnacle_main/framework/constants/size.dart';
 import 'package:pinnacle_main/framework/constants/strings.dart';
 import 'package:pinnacle_main/framework/widgets/components.dart';
+import 'package:pinnacle_main/framework/widgets/navigation_bar/custom_bottom_navigation_bar.dart';
 import 'package:pinnacle_main/framework/widgets/notification_tab.dart';
-import 'package:pinnacle_main/framework/widgets/tool_bar_widgets.dart';
 import 'package:pinnacle_main/framework/widgets/user_widgets.dart';
 import 'package:sizer/sizer.dart';
